 Demo/requirements.txt
pandas
requests
tqdm
matplotlib
wordcloud
streamlit
reportlabs